pyproject.toml
[project]
name = "chat-gemini"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "grandalf>=0.8",
    "langchain-community>=0.3.29",
    "langchain-core>=0.3.75",
    "langchain-google-genai>=2.1.10",
    "python-dotenv>=1.1.1",
]
